developing for abinit [kind of working well now]

+++ emuhelper/abinit/opt.py
@@ -17,43 +17,39 @@ class opt_run:
         self.system = abinit_system(xyz_f)
         self.electrons = abinit_electrons()
         self.ions = abinit_ions()
-        
-        self.electrons.params["ecut"] = 50
-        self.electrons.params["ixc"] = 11 
-        self.electrons.params["occopt"] = 3  # fermi dirac smearing of occupation
-        self.electrons.params["nstep"] = 100
-        self.electrons.params["diemac"] = 2.0
-        self.electrons.params["toldfe"] = "1.0d-6"
-        self.ions.params["ionmov"] = 3
-        self.ions.params["optcell"] = 0
-        self.ions.params["ntime"] = 100
-        #self.ions.params["nctime"] = 1 # for netcdf writing
-        self.ions.params["tolmxf"] = 0 #5.0e-4 # Ha/Bohr
-        self.ions.params["tolmxde"] = "0.0001 eV"
 
-    def gen_input(self, directory="tmp-abinit-opt", inpname="geometric-optimization.in"):
-        if os.path.exists(directory):
-            shutil.rmtree(directory)
-        os.mkdir(directory)
-        os.system("cp *.psp8 %s/" % directory) 
+        self.electrons.basic_setting()
+        self.ions.basic_setting()
 
-        with open(os.path.join(directory, inpname), 'w') as fout:
-            self.electrons.to_in(fout)
-            self.ions.to_in(fout)
-            self.system.to_in(fout)
+    def optimize(self, directory="tmp-abinit-opt", inpname="geometric-optimization.in", mpi="", runopt="gen",
+            electrons={}, kpoints={}, ions={}):
+        if runopt == "gen" or runopt == "genrun":
+            if os.path.exists(directory):
+                shutil.rmtree(directory)
+            os.mkdir(directory)
+            os.system("cp *.psp8 %s/" % directory) 
 
-        with open(os.path.join(directory, inpname.split(".")[0]+".files"), 'w') as fout:
-            fout.write("%s\n" % inpname)
-            fout.write("%s.out\n" % inpname.split(".")[0])
-            fout.write("%si\n" % inpname.split(".")[0])
-            fout.write("%so\n" % inpname.split(".")[0])
-            fout.write("temp\n")
-            for element in self.system.xyz.specie_labels:
-                fout.write("%s\n" % (element + ".psp8"))
-    def run(self, directory="tmp-abinit-opt", inpname="geometric-optimization.in"):
-        os.chdir(directory)
-        os.system("abinit < %s" % inpname.split(".")[0]+".files")
-        os.chdir("../")
+            self.electrons.set_params(electrons)
+            self.electrons.kpoints.set_kpoints(kpoints)
+            self.ions.set_params(ions)
+            with open(os.path.join(directory, inpname), 'w') as fout:
+                self.electrons.to_in(fout)
+                self.ions.to_in(fout)
+                self.system.to_in(fout)
+
+            with open(os.path.join(directory, inpname.split(".")[0]+".files"), 'w') as fout:
+                fout.write("%s\n" % inpname)
+                fout.write("%s.out\n" % inpname.split(".")[0])
+                fout.write("%si\n" % inpname.split(".")[0])
+                fout.write("%so\n" % inpname.split(".")[0])
+                fout.write("temp\n")
+                for element in self.system.xyz.specie_labels:
+                    fout.write("%s\n" % (element + ".psp8"))
+
+        if runopt == "run" or runopt == "genrun":
+            os.chdir(directory)
+            os.system("abinit < %s" % inpname.split(".")[0]+".files")
+            os.chdir("../")
     
     def analysis(self, directory="tmp-abinit-opt", inpname="geometric-optimization.in"):
         pass
